Fix the next Brad testing debacle
I'm waiting for the impeachment vote...

In adding a new bounded buffer exercise, I added a new directory, as
I had with c-ray, to avoid the top-level clutter without realizing
that we had a test named boundedBuffer.chpl in that same directory.
This caused the compiler to want to create an executable named
boundedBuffr as you'd expect, which conflicted with that directory
name.  Oops.  Testing noise ensued.

This commit pushes the old bounded buffer example down a directory
level to avoid the conflict.  In retrospect, maybe I should've
pushed all the old tests down due to their age (I can't remember
the last time we used any of these as hands-on exercises), but I'm
too tired tonight and just wanted to fix testing.

Sorry for the noise everyone.