Fix a race condition in associative array set ops
Consider the following snippet:

```chpl
var dom : domain(int, parSafe=true);
var A : [dom] int;
forall i in 1..100 {
  dom.add(i);
  A[i] = i; // WRONG!
}
```

Adding indices in parallel is fine (provided parSafe is set to true).
Where this program goes wrong is when it attemps to write to an element
while other indices/elements are being added.

``A[i]`` returns a reference to an integer stored in the associative's
array internal array buffer. If an index is added in parallel and the
array is resized, then that reference will now be pointing to invalid
memory. This isn't an issue if the elements are heap allocated.

It's conceivable then that the code could be interleaved like so:
```
   TASK 1           Task 2
-------------    -------------
ref ai = A[i];
                 dom.add(i); // causing a resize
*(ai) = i;
```

This commit updates the internal set operators on associative arrays to
add all the indices before writing to any elements.