RVF fields with record-wrapped types
My initial attempt at RVF-ing fields with record-wrapped types was to
send them across the on-statement inside a bundle, then unpack them on
the other side. It turns out that it's much simpler to just replace any
fields that are references to record-wrapped types with the
record-wrapped type itself. Once doing this, I need to clean up the AST
to fix references that were initialized with the old field reference.

This should eliminate the GETs on non-zero locales in promoted stream.

I also fixed an IWR issue while debugging.