Get hpl.hpcc2012 working again
The problem was that backwardSub was returning a slice of
a local array. That operation is not legal anymore since
arrays are not reference counted. The solution was to
return the local array and then to take the slice of it
at the call site.

While there, I made the printArrays config const actually
do something. This helped me in debugging.

Future work - create a compiler error for returning a slice
of a local array.