Add an OpenMP variant of the task spawning perf tests
Add an OpenMP comparison to the task spawning suite. This is the OpenMP
equivalent of our forall version. We call an extern proc instead of just using
a test.c because this allows the test to work with launchers, use the same
number of tasks, and use the same timers.

On chapcs, the OpenMP version is roughly 7 time faster (~3s vs. ~21s) right now