Replace struct holding the init call I found and a bool with an enum
This enum represents the state the body is in, as determined by what .init()
call was found, if any.  It has four states, only three of which are used at
the moment (the fourth will be used when if statements are appropriately
recognized): when the body is Phase 2 only, when we found a super.init() call,
and when we found a this.init() call.

Thread the use of this enum through getInitCall, phase1Analysis, and
loopAnalysis.  In preparation for a later time when it will be possible to have
arbitrarily many .init() calls be valid (nested if statements, for instance),
I have stopped sending the value along and instead made each Expr make the check
for whether it is a .init() call before performing other analysis on that Expr.