Add compiler errors for returning slice of local array
Code such as

      proc badReturnSliceLocal() {
        var A: [1..100] int;
        return A[1..5];
      }

currently causes core-dump since array slices have no impact on array lifetime.
The variable A is freed before this function returns, but the returned slice
still refers to it. Perhaps array slices should return by-value (ie make a copy
of the elements) in this case but in the mean time it is useful to have a
compiler error instead of a core dump.

I added this check to checkResolved. While there, I moved the check for
returning a reference to a local or constant to checkResolved since as far as I
know it does not need to be in resolution proper.

The new code makes good use of for_SymbolSymExprs which I added in PR #4827.  I
think this represents an improvement over the previous state of affairs:

 * checking for invalid ref returns now happens in checkResolved and
   doesn't confuse resolution's core logic
 * now we get a compilation error for most of the cases that would cause
   a core dump with the new array memory manegement scheme.
   (but note that the simple analysis I implemented is not interprocedural
    and so could be defeated by adding indirection through functions).