RVF should only forward references to const things
I made the naive assumption that RVF should be allowed to forward any
argument with INTENT_CONST_REF. This shouldn't be allowed if the
argument might alias another non-const-ref argument to the on function.

The solution, I believe, is to improve inferConstRefs such that it can
correctly apply FLAG_REF_TO_CONST. RVF will then look for this flag when
deciding whether or not to forward an argument.

The first part of making such a determination is to identify Symbols
with QUAL_VAL and determine if they could have QUAL_CONST_VAL instead.
This is currently done in a very conservative way, only accepting
something as const-val if:
  1) It's passed to a formal with const-ref
  2) It's used in an ADDR_OF where the LHS is a const-ref
  3) It's only used on the LHS of a MOVE/ASSIGN once
In all other cases, I take the safe approach and assume the symbol is
not const.

Now that a number of references and vals have constness info, I can go
about inferring if something is a reference to a const. Here are the
acceptable conditions:
  1) If the symbol is a formal, it is ref-to-const if all corresponding
  actuals are also ref-to-const
  2) If on the LHS of a move/assign...
    2a) and the RHS is a reference, the RHS must also be ref-to-const
    2b) and the RHS is a val, it must be a const-val
  3) It's used in one of a number of primitives that do not change the
  const-ness or address of the reference, as defined by the helper
  function isSafeRefPrimitive.

I also adjusted the re-entrant functions to only change properties on
symbols iff they were the first to be called on a symbol.