Add performance runs over more sizes for the NAS Parallel Benchmark MG
Prior to this change, we were only performing performance runs over size B of
the NAS Parallel Benchmark MG implementation.  Since the smaller sizes are not
very expensive and we run them for the other NAS Parallel Benchmarks, add
performance runs for size S and A as well.

I chose not to add runs for size C yet, as it runs out of memory on one of our
oldest performance testing boxes, and takes more than an hour to complete on
another of our oldest performance testing boxes.  It completes fairly quickly on
chap04, though, so we might want to consider running it everywhere except those
older boxes.

Note that other NPB implementations include a size W in between size S and A.
This implementation does not currently support size W, or sizes D and E.  I took
a brief look at adding size W, but concluded it was more than a
Five-Minute-Fix (tm), so tabled that action for now.