Add --multilocale-only and --max-locales options to start_test (#4846)
Add --multilocale-only and --max-locales options to start_test

[reviewed by @ben-albrecht]

The --multi-locale flag to start_test causes sub_test to run only
tests that specify running on more than 1 locale.  That is, skip any
tests that would be run with -nl 1.  It has no effect when used with
CHPL_COMM=none (a warning is logged).

The --max-locales flag limits the number of locales to use.  This
functionality is already implemented in sub_test by the environment
variable CHPL_TEST_NUM_LOCALES_AVAILABLE, and this commit adds a
start_test flag for it.  The environment variable is still supported,
but will be overridden by --max-locales.