Change to / instead of . to separate build/install subdirs
See email thread "error building Chapel" on chapel-developers from Nov
23, 2016. A user was having trouble building Chapel on a system using
e-cryptfs which has a smaller than usual maximum filename length. This
problem was encountered once before by another user.

This PR adjusts printchplenv to use '/' instead of '.' to separate
the configuration-subdirectory for build/ bin/ and lib/. It makes
minor adjustments to get compiler/Makefile working with this change.
It adjusts runtime/Makefile.help to use printchplenv --simple to generate
a chplconfig file next to libchpl.a.

The bin/ directory does not get a chplconfig file, but the portion
of the configuration used for the compiler can be read either from
the path to the built-for file or its contents. For example:

  bin/darwin/.built-for/darwin/clang/wide-struct/llvm-none/built-for

contains

  darwin/clang/wide-struct/llvm-none