LCALS testing updates
Remove the --dataParMinGranularity=1000 option so it is a more direct
comparison against the OpenMP version.  This unfortunately causes the short
versions to take much more time.

Stop running the vectorizeOnly variant. It is effectively identical to the
serial variant, so is not interesting to run currently.  Remove it from the
perfkeys and graph files.