Commit to save my progress: stop making call to initializer at head of its body.
This results in extraneous extra initialization of the fields.

This commit needs to be cleaned up.  Right now it is not as well organized.

Results in two failures on master (init_before_userdef_initializer.chpl and
bothInitCalls.chpl).  The first failure is due to not providing Phase 1
operations for initializers that only define a Phase 2, the second is due to not
recognizing super and this.init() calls within if statements.