Work around GASNet file manipulation race.
Force a serial make for the build step.  The GASNet build system puts
shell code into config.status that changes the gasnet_config.h file, but
it does so without conflict avoidance and config.status is source'd by
all the conduit builds, among other things.  The result is when the make
is parallel and two conduits are built (near-) simultaneously, we can
end up emptying gasnet_config.h.  The result we've most commonly seen is
a gripe about GASNET_PAGESIZE not being defined.  This problem has been
reported to the GASNet team, but we need to work around it locally until
they can fix it.

While here, I re-deleted the .NOTPARALLEL target, because I'm convinced
we don't need it.