Figure out when we underestimate the heap page size
Addresses a problem with the muxed tasking layer when using GASNet on a Cray XC
system but without a hugepages module loaded.

In that configuration even Hello World would crash.  The problem was that in
that configuration, GASNet arranges for the heap to be on huge pages.  Then
muxed tries to create guard pages, but the mprotect call to do so creates a
guard hugepage (much too large) instead of a guard page. The result is a core
dump when the following code tries to use adjacent memory.

This patch adds logic to chplsys.c to try calling mprotect and accessing memory
in sequence in a way that experimentally determines the page size of the Chapel
heap. This new logic is run from chpl_mem_init.