Solve a couple of issues related to record initializers
First, records don't store a "super" as a field.  We shouldn't expect the first
field to be "super" in the case of records.

Second, while super.init() calls are necessary for the dividing line between
Phase 1 and Phase 2 of the initializer body, the call itself will not resolve
for records.  Record inheritance still requires more thought.  In the case of
"Phase 2"-only initializers, I was inserting a super.init() call that would
then fail to resolve at resolution time.  For the initializers where I inserted
this call, remove the call once we have finished inserting initialization for
those fields.  I added some skeleton code for removing the call when the user
explicitly wrote it in a record initializer (but don't remove it just yet).