Another, less okay, testing update
This is more problematic.  While the initializer in this test is still not
getting analyzed correctly because we don't recognize .init() calls in if
statements, inserting omitted initialization and a super.init() at the
beginning of the body tickles a problem with omitted initialization that I had
not run into yet - namely, that there are still some default initialization
storage issues for fields.  When I try to save the default value for the field
in this test, it is an UnresolvedSymExpr, and later operations in
scopeResolution replace that UnresolvedSymExpr with a SymExpr to the proper
symbol.  That means that 1) the init value stored for this field gets cleaned
up, because it is no longer in the tree, resulting in a segfault when we try
to use it, and 2) we're not storing something that would be usable at normalize
- we'd want to directly insert a SymExpr reference for that field if its
initialization was omitted, so even if the init value wasn't garbage bits, it
would still be the wrong thing to use.

This is going to require a more principled solution - in the past, I have
responded to the initialization stored getting overwritten from under me by
storing the container for the UnresolvedSymExpr, rather than the
UnresolvedSymExpr itself.  However, for this case, the container for the
UnresolvedSymExpr is a call to _createFieldDefault and I don't think that's
something I want to store . . .