Clean up to my normalize fix
Move any support for calling the default constructor to only occur when we
aren't operating on an initializer, including the "constructor for class
requires a generic argument" error message.

While still not as readable as I would like, this makes the operations occurring
more understandable than the quick fix was.