Change the barrier class into a record
Changed the barrier class into a record so that it doesn't require explicit
deletes.

Updated chpldoc comments for a few of the methods.

Documented the 'reset' method in order to expose it.

Updated some tests to remove deletes of barriers.  There may still be more,
but I'll catch them with a round of testing before making this a PR.

Added another barrier to each SPMD LCALS kernel. While removing barrier
deletes, I realized that a barrier was required both before and after
updating the loop index variable, but previously only put one after the
update.