Fix memory leak with push_back
First, pushback-no-leak.chpl is an example that Ben Albrecht provided.
This program was leaking memory because the array-as-vector support
creates extra space at the end of the array to make future additions not
need to reallocate. The problem was that these array elements were being
initialized but never freed.

Fixing this issue amounts to adjusting dsiDestroyArr to figure out the
size of the allocated array by looking at dataAllocRange in addition to
dom.dsiNumIndices. However, in order to have the current dataAllocRange
in the call to dsiDestroyArr in the call to dsiReallocate inside e.g.
push_back, I needed to also modify push_back and the other array-as-vec
routines to only set dataAllocRange after the reallocation has completed.
To do so, I introduced a temporay variable called nextAllocRange.