Significantly rework omitted field initialization
At Mike's suggestion, stop creating the default constructor when we have
defined an initializer for that type.  As a result, this means that scope
resolution no longer removes the field initialization and type information from
the fields to place them in the default constructor.  This meant that I could
use the field initialization and type information directly from the fields
themselves when inserting omitted field initialization, instead of having to
save that information in a brittle and error-prone form.

Because of this, I was able to toss a lot of fragile code from initializerRules.

This change resulted in fixing two of the tests I had just broken with my
previous merge (constAndParamInHeader and bothInitCalls).  It also resulted in
four new failures, which I still need to investigate.  Saving my progress,
though, because this is a significant change and I didn't want to lose it while
fixing the minor problems.