Work around broken osx+gnu jemalloc support in our makefiles
This reverts #4878, which worked around osx+gnu support by modifying the
jemalloc source, and replaces it with a workaround in our jemalloc makefile.

We now tell jemalloc's configure step that __builtin_unreachable is broken on
osx+gnu instead of letting it try to figure out it for itself since the check
for it is currently broken.

See https://github.com/jemalloc/jemalloc/pull/504 for more info