Upgrade to jemalloc 4.4.0
jemalloc 4.4.0 includes some optimizations and bug fixes.

Optimizations that we care about:
 - "Add extent serial numbers and use them where appropriate as a sort key that
   is higher priority than address, so that the allocation policy prefers older
   extents. This tends to improve locality (decrease fragmentation) when memory
   grows downward."

This actually causes a slight slowdown (4-5%) for binary-trees, but this is
likely an artifact of more purging occurring (leading to less fragmentation) so
it's actually a good thing. See https://github.com/jemalloc/jemalloc/issues/147
for more info.

For the full changelog see:
 - https://github.com/jemalloc/jemalloc/releases/4.4.0