inferConst should consider _retArg to be a single def
The compiler currently transforms functions that return certain types
(e.g., tuples, records) such that they will return through a reference
formal instead of a PRIM_RETURN. Such formals should not be considered
to have the const-ref intent, as they will modify the thing they are
pointing to.

The inferConst function, however, can consider this to be a single 'def'
of a value. This may allow for the value to become a const-val, which
can allow RVF to be more effective.