Fix extra -O0 for PrgEnv-cray
PR #1929 caused PrgEnv-cray compiles to always include -O0
(and possibly also -O3 if fast is used). @ronawho pointed out
that the extra -O0 is having an impact on compilation.

This PR fixes the problem by adjusting Makefile.cray-prgenv-cray to put -O0 in
NO_OPT_CFLAGS instead of adjusting CFLAGS directly.  Then, when
runtime/etc/Makefile.include computes MAKE_COMP_GEN_CFLAGS, it adds in
NO_OPT_CFLAGS if optimizations are disabled (and as before it adds in
OPT_CFLAGS if optimizations are enabled).

Verified that PrgEnv-cray and PrgEnv-gnu behave as expected
by manually inspecting --print-commands with all combinations
of --fast and -g.