Update some failing futures due to my initializers changes from yesterday
I altered the analysis of initializers to insert Phase 1 initialization in
initializers that were otherwise Phase 2 only bodies.  This caused
test/types/records/ferguson/copy-init/new-init-generic.chpl to fail to match its
.bad file, because the initialization of the generic var field was not
explicitly in Phase 1 and so was viewed as omitted during that Phase.  I updated
the test to adjust for this.  My handling of "super" for records in this case
is not correct, so this update also impacted the .bad file (which would have
needed to be changed anyways because we do not correctly grab the initialization
and type information for generic fields yet anyways).

I also altered the handling of initializers to no longer insert a call to the
default constructor, changing the error message for constructors that elided
necessary generic arguments to indicate that they only impacted constructors.
This caused test/classes/vass/nested-class-with-user-defined-constructor-1.chpl
and test/types/records/ferguson/copy-init/old-ctor-generic.chpl to no longer
match their .bad files.