Change build_record_init_function for initializers
Now that we don't generate a default constructor for types which define
initializers, that function was running into a problem because it was trying to
use what was a NULL pointer in that case.  I altered it to only call out to the
no argument version, and if one isn't resolvable, then we'll encounter a
resolution error and that is okay.

Note that this doesn't handle generics right.  The old version gets the generic
fields from the default constructor, but we can't rely on that.  Punt for now.