Testing updates due to this change
Remove the future for constAndParamInHeader and bothInitCalls, as they pass
again (yay!).  Add a future to init_before_userdef_initializer and to
phase1/dependence (due to not inserting "this." on accesses of earlier fields).

I have not resolved the error with badDependence2 yet, I am checking to see if
the fix is easy or if I should just live with it (because while I understand
getting that message, I'm not sure why I started getting that message *now*).

Also not added is an alteration to nested-destructor, because I have an action
item to ask Michael whether that change is reasonable (independent of the work
related to it that I already know I need to do)