Add future to badDependence2, which now fails with "unable to resolve type"
badDependence2 tracks what happens when a field refers to another field that
has not been initialized yet.  It used to recognize a "use before def" had
occured on that field, but that appears to be because it didn't recognize that
the symbol refered to was a field at all.  Now it does recognize that (and
normalize's "use before def" check doesn't distinguish between these actions
that happen in initializers/constructors and ones that happen else to fields,
choosing to just assume that the field has been appropriately updated elsewhere)
and will skip that check, leading to problems when trying to determine the
type for the field later (because we won't have seen its first def, and so
won't have propogated the type to it).

This is obviously wrong, but is not in the scope of this PR.