Changing my fasta timer from default to high-precision
On the shootout machine, I'm seeing dramatically different timings
for command-line time vs. the h.p. timer.  I haven't seen the same
on my Mac or bradc-lnx (in my home directory), but in any case, it
doesn't seem like a good idea to be using different timers on
different versions.