Adjust verify on AggregateTypes
After my change yesterday (PR #4980), verification started failing for my
initializer tests.  This is because --verify expects all AggregateTypes to have
their defaultInitializer field defined, but we've changed the expectation for
initializers so that if any initializer is defined, the default constructor or
initializer will not be generated.  Update the check to recognize that exception
to its rule.