Add --ieee-float to HPCC FFT test
A reasonable alternative would be to change the tolerance in the
test. I tried quickly looking at the HPCC challenge reference version
but it wasn't immediately obvious how the error checking corresponded.
In particular , the reference version (1.5.0) uses

    if (maxErr / log(n) / deps < params->test.thrsh) failure = 0;

where params->test.thrsh is 16 by default, and deps may be the
machine precision. Our version does not seem to have this division
by log(n) and I do not know why.