Try a numa-sensitive version of fasta
This is a variation on fasta-blc-numa.chpl which strives to be
sensitive to NUMA placement by jumping through some hoops in hopes of
damping down some of the noise that seems to take place with the
fasta-blc version on systems with multiple numa domains (e.g.,
chapcs07).  Specifically, what seems to be happening is that when we
create 3 tasks, they are dealt out to the numa domains cyclically such
that they are spread across numa domains; yet all of the data that
they want to touch should be local to the same numa domain.  This
version works around that issue by creating 2*numTasks tasks and then
only having the even ones do any work (where '2' is stored in a param
called 'numSockets)', causing them all to be mapped to the same
domain.  From what I've seen, this seems to stabilize timings.

I've also added an additional coforall loop after the main one to
create an additional 'here.numTasks - 2*numTasks' additional tasks,
just in case creating a number of tasks equal to the number of sockets
matters.  It seems like it shouldn't and yet, in some of my early
runs, it did seem to have a positive impact (though I'm not seeing
that now that I've created a separate test for it...).