Make sure # of tasks doesn't exceed here.maxTaskPar
The previous version of this created deadlock on my low core count
system due to the busy-waiting.  Here, I've made sure that the
maximum total # of tasks doesn't exceed here.maxTaskPar to avoid
this which seems to work (but may use fewer than 3 tasks on a
chip with fewer than six sockets, obviously.