Tightened up the heuristics further to avoid loss
In the previous version, the number of tasks would be less than 3 even
on a single-socket 4-core processor that could handle all 3 tasks
because there weren't enough cores for 6 tasks where we throw half of
them away.  Here, I beef up the heuristic slightly to either use k and
2*k tasks if here.maxTaskPar is >= 2*k; if it's not, I use k and k
tasks and don't do the "deal out every other task" strategy
(essentially, falling back on an assumtion that we're on a single
socket node).  This should avoid throwing performance away on
single-socket nodes.