Move test page to the end of the heap
This PR resolves a problem with CHPL_COMM=ugni where allocating and using
a large amount of memory could result in some PUTs to fail. This issue
was introduced in PR #4931.

In order to be sure we correctly know the page size of the Chapel heap,
we did experiments with mprotect in computeHeapPageSizeByGuessing. Before
this PR, those experiments used memory in the middle of the heap. That
has the unfortunate side effect of splitting the heap in the middle as
far as the linux kernel is concerned - which complicates memory
registration.

This PR changes the logic in computeHeapPageSizeByGuessing to use memory
at the end of the heap instead. This memory is less likely to be used in
practice. (However if it is used, there is still something to fix within
CHPL_COMM=ugni).

While working on this PR, I noticed that the experiment with mprotect has
an impact on the "first-touch" algorithm for moving memory between NUMA
nodes. That is because it sets one byte of memory in addition to
mprotecting. This effect too should be minimized by working with the end
of the heap.