Fix race condition in counting number of PUs and begin refactoring locale models (#4999)
There was a race condition in initializing the root locale, caused by each node writing to the number of PUs simultaneously. This change removes the race condition by accumulating the values into atomic variables.

To reduce the amount of duplicated code between different locale models, the new code is placed in a common helper file. This is the beginning of a larger code restructuring to make it easier to implement new locale models such as the one for KNL.

A few expected test results had to be adjusted due to the new module being added. The code has been tested with the flat locale model on Linux and Mac, and the flat and numa locale models on Crays, including KNL.

Reviewed by @mppf  -- thanks.