Improve task creation speed for "bounded" coforalls
Improve task creation speed for "bounded" local coforalls by limiting how many
times we touch the EndCount and runningTasks atomics.

Previously coforalls were turned into something roughly like:

```chapel
coforall i in 1..10 { body(); }
```

```chapel
for i in 1..10 {
  _upEndCount(_coforallCount);
  chpl_taskListAddCoStmt(body());
}
_waitEndCount(_coforallCount);
```

where `_upEndCount` bumped the atomics here.runningTasks, EndCount.i, and the
non atomic EndCount.taskCnt. Modifying the atomics for each iteration is
actually quite expensive so this changes the code to do a single upEndCount for
a coforall when we can determine the number of tasks the coforall will create
before executing it. Currently that check is if the coforall is over a
boundedRange, a domain, or an array.

We now create something like:

```chapel
var tmpIter = 1..10;
if isBoundRange(tmpIter) || isDomain(tmpIter) || isArray(tmpIter) {
  var numTasks = tmpIter.size;
  _upEndCount(_coforallCount, numTasks);
  for i in tmpIter {
    chpl_taskListAddCoStmt(body());
  }
  _waitEndCount(_coforallCount, numTasks);
} else  {
  // old style coforall
}
```

which allows us to only bump the EndCount and here.runningTaskCount once, and
allows us to avoid using taskCnt since we're now passing the number of tasks
created directly to waitEndCount.

Note that this does effectively disable the anonymous range/domain optimization
for coforalls since we're de-anonymizing the iterator. That's unfortunate, but
the performance benefit of not manipulating atomics far outweighs the cost of
constructing a range/domain.

This has a pretty significant performance improvement for several benchmarks:

 - 50% speedup for forall/coforall task microbenchmarks (within 20% of gomp now)
 - 25-35% speedup for short parallel LCALS
 - 10% speedup for long LCALS, lulesh, miniMD, NPB FT, and a few others

Some future follow-ons to this include:
 - apply this optimization to coforall+on stmts
 - come up with a better way to detect "bounded" coforalls instead of just
   checking for boundRange/domain/array
 - figure out a way to still apply the anonymous range/domain optimization
 - have "bounded" coforalls call into an optimized runtime routine that spawns
   multiple tasks at once instead of calling chpl_taskListAddCoStmt once per
   iteration.