Moved lowering of PRIM_DELETE to resolve(); replaced compiler-generated AST with a Chapel function, chpl__delete().
This is motivated by the upcoming ForallStmt node that will be lowered
during resolve(). It requires "delete" of iteration classes, so
inability of generate PRIM_DELETE during resolution would be a hindrance.

* Renamed "~chpl_destroy" to "chpl__deinit". This is so that it can be
invoked from Chapel code, specifically from chpl__delete().

* PRIM_DELETE now lowers directly to a call to chpl__delete() in preFold().
Alternatively we could generate a call to chpl__delete() right in the parser
and so dispense with PRIM_DELETE. I did not do that so that 'delete'
remains as a primitive operation, in case we want to do something about it
before resolution.

* Added a check to ensure that user-defined destructors have parens.
Otherwise the call arg.chpl__deinit() in chpl__delete() does not resolve.
Added missing parens in a couple of tests.

* Reduced comm counts (gets) in a couple of tests.

* Potential future improvement - better error message when 'delete' is
invoked on a non-class.
  Right now the cases of records and _ddata are caught explicitly.
However, the cases like delete of an integer result in a "unresolved
call to chpl__deinit()" error, with a ton of chpl__deinit candidates
printed out.
  An issue here is that the isClass() function in our modules returns
false for "extern" classes, whereas we DO delete those in our tests.
  This all is unaffected by this PR.