Make libnuma available on Crays when needed (#5016)
The memkind library requires libnuma, and there is no static version of either one of them.

On Crays, when the cray-memkind module is loaded (which also sets the default link mode to dynamic), we get libnuma whether we want it or not, so hwloc should take advantage of it. This change does so for that case.

On Crays, when a non-flat locale model is used, we need hwloc to have access to libnuma, otherwise we will be unable to specify from which numa node a memory allocation should occur. This change enables libnuma for that case as well.

This was tested on Mac, Linux, and Crays, including KNL with and without memkind, and with and without a flat locale model.

Extensive discussions with @ronawho and reviewed by @gbtitus -- thanks!