Accept paren-less destructors + cleanup
* Do allow paren-less destructors. See #5019 discussion
and also (less contents) the thread "parens required for deinitializers"
on the internal list.

* For that, have the compiler add parens internally if the user didn't.
This is implemented by removing FLAG_NO_PARENS.

* Reverted corresponding changes to tests.

* Move the INT_ASSERT check that a yield is always within an iterator
from normalize() to check_checkParsed(). Note that if the user made
an error, checkParsed issues USR_FATAL_CONT, so the check is violated
and the compiler has not aborted yet. Force exitIfFatalErrorsEncountered()
before running that check.