Change loop idiom, rename some things
This changes the loop that chunks up the 1..n iteration spcae by
* changing the low bound rather than using an 'align' keyword
* changing the loop to use 0-based indexing

The use of 0-based indexing in particular enables some other
nice changes such as some +/- 1 math and removing the initializations
of the guard arrays to 1.

While here, I also noticed that I had two loops that both used
'i' as their index variable, so renamed the inner loops to 'j'
and 'k'.  I also changed a variable name to 'nid' for brevity
and arguably clarity.