Continue reducing diffs between bradc and release versions of fasta
...with one key difference:  The release version showed that it
performed better when the atomics were explicitly initialized
which is counterintuitive given that it's a parallel initialization
of a 3-element array.  My best guess is that that operation somehow
keeps the tasking layer "warmer" in a way that benefits the
benchmark.

For the release version for now, though, I'm leaving it in its
simpler / more minimal state since that is how I'd prefer to
write the code.  We can see how performance varies across the
systems upon checking this in.